backend/requirements.txt
Django==2.2.20
django-filter==21.1
djangorestframework==3.12.4
drf-extra-fields==3.1.1
pytz==2022.6
sqlparse==0.4.3
